requirements.txt
streamlit
pandas
numpy
scikit-learn==1.6.1
plotly.express
seaborn
matplotlib.pyplot
tensorflow
